requirements-qat.txt
--extra-index-url https://download.pytorch.org/whl/cu128

torch==2.10.0+cu128
torchvision==0.25.0+cu128

nvidia-modelopt==0.41.0

numpy==1.26.4
pandas==2.2.3
matplotlib==3.10.0
pillow==11.0.0
tqdm==4.67.1
PyYAML==6.0.2
psutil==7.2.2

jupyter==1.1.1
jupyterlab==4.5.6
notebook==7.5.5
ipykernel==7.2.0
ipywidgets==8.1.8
